Refactored delegateRow utils
