+++ src/modules/dpos/validator/components/delegatesTable/delegateRow.js
@@ -5,6 +5,7 @@ import { useDispatch } from 'react-redux';
 import routes from '@screens/router/routes';
 import { useTheme } from 'src/theme/Theme';
 import { addedToWatchList, removedFromWatchList } from '@common/store/actions';
+import getForgingTime from '../../utils/getForgingTime';
 import styles from '../delegatesMonitorView/delegates.css';
 import {
   DelegateWeight,
@@ -14,19 +15,6 @@ import {
   ForgingTime,
   DelegateRank,
 } from './dataColumns';
-
-const getForgingTime = (time) => {
-  if (!time) return '-';
-  const diff = time - Math.floor((new Date()).getTime() / 1000);
-  if (Math.abs(diff) < 9) return 'now';
-  const absTime = Math.abs(diff);
-  const minutes = absTime / 60 >= 1 ? `${Math.floor(absTime / 60)}m ` : '';
-  const seconds = absTime % 60 >= 1 ? `${absTime % 60}s` : '';
-  if (diff > 0) {
-    return `in ${minutes}${seconds}`;
-  }
-  return `${minutes}${seconds} ago`;
-};
 
 const DelegateRow = ({
   data, className, t, activeTab, watchList, setActiveTab, blocks,
